D-Bus Testing: fixed TestDBusServerPresence + TestDBusSession.testSecondSession
TestDBusSession.testSecondSession failed when TestDBusPresence tests
had been run before. The reason is apparently the mock Connman object
and its calls to loop.quit(): that causes the testSecondSession test
to stop before it has seen all the expected reasons for quitting the
main loop.

This commit changes TestDBusServerPresence so that the Connman object
is added and removed as part of setUp() and testDown(). This seems to fix
the problem.

Connman.GetProperties() now also returns something valid in the final
else clause. Previously Python recorded a lot of "'None' not iterable"
errors when Connman.GetProperties() was called more often than
expected and returned None.