D-Bus testing: fixed TestSessionAPIsDummy.testCheckSourceInvalidType
The test was meant to check the error triggered by setting an
invalid backend value. Instead it checked the usability of such
a source and thus duplicated the (badly named) testCheckSourceNoType.

testCheckSourceInvalidType itself failed to pass when SyncEvolution
was compiled with modules, because then the "apple-contacts" backend
wasn't installed and SetConfig() failed with an unexpected error.

Now the test triggers that error in call cases with "backend = no-such-backend"
and checks that the right error is reported.