nightly testing: join multiple CATEGORIES
As per RFC 2445, CATEGORIES:1,2 and CATEGORIES:1\nCATEGORIES:2 is
same. We need to pick one normal form. This commit ensures that all
categories are listed in a single CATEGORIES properties. This was both
easier to implement (splitting at a comma while not splitting at a \,
is tricky) and leads to a shorter normal form (less lines).

The reason for making this change now is that with Evolution 2.32.2
and libical 0.44-3, categories end up being stored with one entry per
CATEGORIES property. That by itself is okay and thus our tests should
pass, which they don't without this change to synccompare.

The only problem is that Evolution doesn't seem to handle it (breaks
setting categories in the UI even without syncing). That's something which
needs to be fixed in Evolution.