SyncConfig: fixed compiler problem with C++ initialization order (BMC #19464)
g++ 4.6 and ld 2.21.52.20110707 (Debian Unstable) led to a different
order of global instance construction:
1. WebDAV constructor calls SyncConfig::getRegistry()
2. getRegistry() adds the (uninitialized!) property instances
   and modifies them
3. SyncConfig.cpp instances are initializes, which resets
   some of the values modified by getRegistry()

The result was that, for example, the "defaultPeer" property was
treated like an unshared property and written into the wrong config
file.

The assumption that variables in a compilation unit are initialized
before methods in that unit can be called is not based on anything in
the C++ standard. Therefore this commit rewrites the code so that
properties are not added/updated inside the getRegistry()
methods. Instead this is done in separate classes which (and that is
guaranteed by the C++ standard) are constructed after the properties
defined earlier in the compilation unit.