nightly testing: must pre-start gnome-keyring-daemon
On Ubuntu Lucid, GNOME keyring is not started when the GNOME keyring
library asks for the org.freedesktop.secrets service. This
broke testing of syncevo-dbus-server.

This commit solves the problem by explicitly starting and stopping
that daemon as part of creating the D-Bus session. This is similar
to what happens on Lucid when a GNOME session is started.