testing: relax requirements for ActiveSync backend
In an incremental sync, the ActiveSync backend is only able to provide
data for updated or new items. This is sufficient, other items will
not be read during such a sync and item access doesn't use incremental
sync mode.

Therefore this commit relaxes the testChanges test by removing the
reading of an unmodified item.