testing: report source initialization errors inside test*() methods
TestingSyncSourcePtr used to catch all exceptions thrown while
constructing and opening the source. This had the negative effect that
errors logged contained the source code line inside
TestingSyncSourcePtr instead of the line where the test itself asks
for the source, which is much more important for debugging the
problem.

This commit thus removes the CPPUnit assertion which catches
exceptions and adds them in all locations inside test*() methods
which, directly or indirectly, create sources.

It is not essential to find all of these places. CPPUnit will report
uncaught exceptions in a test without including a source code
location, which is as good (or bad) as the information reported
previously.