autotools: more strict compatibility check (part of BMC #21284)
Some unintended usage of libical slipped into the 1.1.99.5
syncevolution.org binaries, breaking the syncwebdav.so. It wasn't
linked against libical, but depended on some of its methods.

This commit strenghtens the installcheck-local test. First, it checks
for undefined symbols in the blacklisted libraries (detects the
syncwebdav.so case). Second, it also checks for the blacklisted
libraries (might be linked unnecessarily, for example when forgetting
to add -Wl,--as-needed). Third, it checks syncevolution (relevant when
linking everything statically). Finally, it does the check on the
installed files (should have been like that all along).