Google Calendar testing: ignore UTC -> timezone conversion
Google Calendar converts UTC time stamps into the time zone of the
calendar as set when the event is stored in the calendar. This changes
the semantic of the event slight, in particular for recurring events,
but is acceptable and is done correctly for the event's point in time
by Google, so let's accept this change.

This commit adds the Europe/Berlin time zone to the test data used for
CLIENT_TEST_SERVER=googlecalendar Client::Sync::eds_event
so that the change above no longer occurs in the testing and thus doesn't
break the test.

Because a time zone is set explicitly and preserved by Google, the Google
Calendar time zone setting doesn't matter for testing.