Google Calendar testing: avoid "custom time zone" test
Google Calendar recognizes the "custom US time zone" despite using
different names for TZID and TZNAME. Probably uses the offsets. When
it recognizes the zone, it replaces it with its own internal zone,
which breaks the test.

That's okay, so let's stop using this particular test event to get the
test to pass.