testing: fixed eds_memo update tests
Mangling the VJOURNAL test data by prepending "U " to the DESCRIPTION
caused problems because the first line has to match the SUMMARY (in
plain text, only the body is sent and SUMMARY is taken from the first
line).

This commit fixes these problems by also inserting the "U " into the
SUMMARY of memos.