D-Bus testing: removed side effect of TestDBusServerPresence tests
The TestDBusServerPresence tests have to invoke syncevo-dbus-server with
DBUS_TEST_CONNMAN set in the environment, to get it to use the mock
implementation on the session D-Bus.

They did this by modifying os.environment *permanently*. That had the
effect that all later tests also ended up using the mock ConnMan. This
made testing confusing because tests failed only when run in combination
with the presence tests (fixed previously by improving the mock ConnMan
implementation).

This commit addresses the root cause by putting the env variable only
temporarily into the environment of those tests which really need it.