Memotoo testing: strengthen PHOTO comparison
The PHOTO property was removed completely for Memotoo because the
literal comparison used to fail due to transcoding the images. Now
synccompare keeps the PHOTO property and only removes the data itself.

This was done in order to test PHOTO;VALUE=uri. It turned out that
Memotoo does not support those.