vCard: support VALUE parameter, fixed TYPE (part of BMC #19661)
The VALUE parameter was ignored completely and thus got lost in the
Synthesis engine. Added it as a string.

The TYPE parameter was an incomplete enum. Better allow arbitrary
strings.

Both parameters must be copied together with the PHOTO data they
belong to. In combination with merge="fillempty" this can be
problematic: if one of the parameters is empty, it may be overwritten
although the PHOTO data is not copied.

This problem is solved by ensuring that the internal field list never
has empty PHOTO_TYPE/VALUE fields. This is done by setting "binary"
resp. "unknown" when importing contact data from a peer or the local
Evolution backend and removing them again before sending to a peer or
storing.

The same change needs to be done with other backends. It is not made
mandatory because some backends (like file) might want to store these
values explicitly.

Because it is unknown which peers support VALUE=uri, only the
CLIENT_TEST_SERVER=syncevolution testItem testcases contain a contact
with such a PHOTO.