server: Get bluetooth phone vendor and model from Device ID profile (BMC #736)
In the past we have relied on the user-modifiable device name to be
the fingerprint for matching a phone to a template which is unreliable.

This commit changes this in the cases where the phone supports the
Device ID profile (DIP). If support DIP is detected then we extract
the vendor and product id's from the PnPInformation service
record. Then, we attempt to associate the ids with a product and
vendor name by using a newly added lookup table.

The preference for which value to use as the fingerprint is as
follows:
1) If the product is found in the lookup table we use
   that. (The product lookup table is manually updated and currently
   only has a few entries.)
2) Otherwise, we use the vendor name. (All vendors should be in the
   lookup table, so this should always be used if the DIP is supported.
3) If the DIP is not supported by this device we fall back to using
   the user-modifiable device name as before.

Additionaly, because the user-modifiable device name is still
desirable to expose through the dbus interface, it is now made
avaiable through the peerName property. This is intended to be used in
GUIs, for example.