WebDAV: fixed item operations without configuration (BMC #22164)
Item operations like --print-items failed if the configuration didn't exist:
[ERROR] : virtual read-only configuration node, cannot write property webDAVCredentialsOkay = 1

This was caused by trying to write the webDAVCredentialsOkay property
into a temporary, read-only configuration. The fix is to check for
read-only configs before attempting to use the property.

Alternatively it would have been possible to catch exceptions, but it
is not obvious which errors can be ignored.