CalDAV: fixed incorrect change tracking causing "event not found" (BMC #22329)
During an incremental sync, when an unmodified meeting series on the CalDAV
server had to be extended (= adding a detached recurrence), the operation failed
with "event not found".

The root cause was a bug in updateAllSubItems(): due to a copy-and-paste bug,
it cleared the cache instead of adding the modified items to it. Therefore
unmodified items, added to the cache earlier, where not found later on.

Normally they weren't needed. The exception is reading in preparation
for adding a detached recurrences.

Fixed by removing the m_cache.clear(). That is valid in this case even
if the operation is repeated, because adding already read items will
simply overwrite them.

Also added some debug logging which helped to track this down.