testing: added Client::Sync::testDeleteBothSides (for BMC #22327)
Trigger "item already deleted" in syncing, check that syncing
completes and leaves no items around.

Somehow the problem in BMC #22327 must have been different: when
running the test, the Synthesis engine recognizes that the item
was removed on both sides and logs:

So this test doesn't really cover BMC 22327. But it is worthwile
to run it anyway.