testing: optionally ignore leaks in forked processes
Improved valgrindcheck.sh so that if VALGRIND_LEAK_CHECK_ONLY_FIRST
is set, memory leaks in forked processes are filtered out of the
valgrind log file.

Also improved the detection whether an error was reported. This is
the case if and only if a suppression was created. This allows
running valgrind with normal output, no need for --quiet anymore.