dbus-session.sh: optionally start EDS under valgrind
If DBUS_SESSION_SH_EDS_BASE is set to a directory holding
e-calendar/addressbook-factory and the main command includes
valgind or valgrindcheck.sh, then the same command is also
applied to the starting of these two EDS daemons. The total
result of dbus-session.sh then is the error code from the main
command (if not 0) or error codes from either of the two daemons.

Killing via SIGINT and SIGTERM is not considered an error, because
that is how the two daemons have to be terminated by dbus-session.sh.
SIGKILL is an error because it implies that the daemons did not
shut down properly or quickly enough.