valgrindcheck.sh: added signal handling and improved result checking
SIGINT/TERM are now caught by the script and only cause it to forward
those signals to the started valgrind instance. This is necessary
when running EDS under dbus-session.sh, which needs to send these
signals to trigger the daemon shutdown.

In that case, valgrind needs more time for leak checking. Increased
delay before sending SIGKILL.

Be careful to not override the programs return code with the 100 =
"valgrind problems found" status unless the program completed normally.
SIGINT/TERM are considered normal termination because EDS daemons
have to be shut down like that.