testing: better Client::Sync test overview
First, the list of tests was incomplete because there is no single
source which has all tests enabled. Must create a union of all enabled
tests. Do it so that non-standard tests are listed last.

Second, don't break that order when creating the HTML output. Order as
run is better than alphabetical order.