testing: verify that auto sync triggers notifications
This is tested as part of the existing auto sync test which produces
notifications. It would be nice to monitor the method calls via
Python, but it doesn't seem to be possible to watch method calls. So
instead we check the dbus-monitor output after the test has completed.

What makes this a bit harder is that depending on how test-dbus.py is
started, there may or may not be a real notification daemon
running. If there is, then the notifications triggered by testing will
show up on the tester's desktop. If there is not, then
syncevo-dbus-daemon will not send the expected notifications.

To ensure that testing of notifications still works in the nightly testing
where X is not running and thus GNOME notification daemon doesn't work,
a fake implementation of the org.freedesktop.notifications API is
provided and started as part of the test-dbus.py script in a forked
process.

The stand-alone version of the fake notification daemon is also available,
but would be harder to start (PATH must be set, etc.), so it is not used.