CalDAV + Google: further workarounds for linked items
Sometimes a PUT of two linked events updates one of them on the server
(visible in modified SEQUENCE and LAST-MODIFIED values) and then
fails with 409 because, presumably, the other item now has
too low SEQUENCE/LAST-MODIFIED/DTSTAMP values.

This looks like a bug in the Google CalDAV implementation.

As a workaround, the PUT is retried once with increased
SEQUENCE/LAST-MODIFIED/DTSTAMP values.