testing: introduced ClientTestConfig::sourceLUIDsAreVolatile
Necessary for LinkedItem tests when applied to ActiveSync. In that
case, items are renumbered as 1:x with x = 1, 2, ... for each clients
when a sync anchor is assigned to it, and therefore the LUID
comparisons/searches fail although everything is okay.
(cherry picked from commit e45e308806524bada86d57353e24714f06edc331)