ActiveSync: set "sourceLUIDsAreVolatile"
Client::Source::eas_event::TestLinkedItems failed because it made the
assumption that LUIDs are the same for all clients. With ActiveSync
they are not (get renumbered in the initial sync), so this particular
aspect of the tests has to be disabled.

activesyncd commit ID:
fc5994fe8cc9a9d253501527a98f49f4eea8e445