testing: LinkedItems_1 with "exchange" server: recurring all-day event
RFC 5545 and RFC 2445 both specify that RECURRENCE-ID must be
VALUE=DATE if the recurring component uses that.

I wrote a variation of the LinkedItems tests which are meant cover
this. When run with CLIENT_TEST_SERVER=exchange, LinkedItems_1 will
use a weekly recurring all-day event as parent with one exception in
the second week.