Add script to query for the bluetooth device id
We currently only have a few devices added to our bluetooth product
lookup table.

This script is intended to allow user to query the information from
device they own and instructs them to send it to a syncevolution
devoloper (me for the time being). The goal is to crowd-source sa much
phone data as we can.