D-Bus server: send hardware info in new "hardwareName" read-only property
The previously suggested (and implemented) approach of reusing the
peerName property to convey hardware information didn't pass the smell
test (too complex to explain that the same property has different
meanings depending on the context).

So now a new "hardwareName" property is used instead. Also renamed
the TemplateDescription member accordingly, updated the API docs
and implemented the "hardwareName only set if known" part of it.