testing: extended detached recurrence test cases (triggers BMC #22940)
The parent with single detached recurrence was too simplistic. More
than one detached recurrence is necessary to trigger a problem in the
EDS calendar backend.

Because the "parent + single detached recurrence" may also be a relevant
special case, it is also still covered.

Finally, importing multiple detached recurrences without parent also
needs to be covered.

Missing is importing of a single detached recurrence. This should
be covered by importing the first of the two detached recurrences.