WebDAV: ignore props of a found collection
When a collection was identified as one we searched for, then don't
look at its properties. If the search started with the principal, then
we are not going to learn more from the collection's properties. If
the search starts with a specific collection, then the search will stop
after reporting that collection.

Not listing further collections when searching for all databases is
probably what the user wants. Otherwise he should have used the
principal or simple the host name.