SyncSource: allow StartDataRead/beginSync() to force a slow sync
This is needed for ActiveSync, which can detect inside beginSync()
that the sync key became invalid and then has to fall back to a slow
sync.

It is done by asking the engine to call StartDataRead earlier (via
<plugin_earlystartdataread>) and documenting the special 508 error
code. Depends on support for both in libsynthesis, thus the bumped
version number of it.

Normally SyncEvolution is meant to never throw an exception unless
something fatal happens, so it would have been better to add a return
code to beginSync(). But that is a fairly intrusive change, so instead
the simpler "use exception" solution is used. The change can still be
made later on.