D-Bus API: allow setting different configs in the same session
Forcing D-Bus API users to start a new session for each config that
they want to modify adds considerable complexity to clients, in
particular for local sync.

Relaxing that requirements is easy because the current implementation
already only runs one session at a time. It's also only a minor
restriction for future improvements: sessions which request access to
all configs cannot run in parallel with other sessions.

Thus this patch adds Session.SetNamedConfig() and (for the sake of
completeness) Session.GetNamedConfig().

Temporary config changes via SetNamedConfig() are harder to implement
and not currently supported. They are part of the API though, just in
case.