Memotoo testing: supports ADR country
Found out that Memotoo supports the country field, albeit only if the
string is recognized by Memotoo. That wasn't the case for some of the
test data.

Adapted the test data to use countries that Memotoo knows (Germany,
France, with the English names). Also updated synccompare.pl to
reflect the current set of lost properties and removed the Memotoo ADR
simplification.